no really balancer works ok now

+++ dev/test.rb
@@ -55,7 +55,7 @@ def get(url, opt = {})
   Typhoeus.get(url, followlocation: true)
 end
 def get_until(url, fin, opt = {})
-  max_wait = opt[:max_wait] || 10
+  max_wait = opt[:max_wait] || 3
   neg = false
   if fin[0]=="!"
     fin = fin[1..fin.length]
@@ -65,21 +65,29 @@ def get_until(url, fin, opt = {})
   start = Time.now
   while true do
     resp = get(url, opt)
-    puts resp.return_code
     if neg then
       break if resp.return_code != fin
     else
       break if resp.return_code == fin
     end
-    if Time.now - t > max_wait
+    if Time.now - start > max_wait
       raise "failed to get #{url} after #{max_wait} sec."
     end
-    sleep opt["retry_time"] || 0.2
+    sleep opt[:retry_time] || 0.2
   end
   return resp
 end
+def get_repeat(url, reps=100, opt={})
+  codes = {}
+  reps.times do
+    resp = get url, opt
+    codes[resp.return_code]=(codes[resp.return_code] || 0) + 1
+  end
+  return codes
+end
 
 class Nginx
+  @@pid = nil
   def self.start(opt = "")
     @@pid = Process.spawn("./nginx.sh #{opt}")
     get_until "http://127.0.0.1:8082/ready", :ok
@@ -89,13 +97,13 @@ class Nginx
     @@pid
   end
   def self.stop
-    Process.kill "TERM", @@pid
-    print "stop nginx"
+    Process.kill "TERM", @@pid if @@pid
     get_until "http://127.0.0.1:8082/ready", :couldnt_connect
   end
 end
 
-Nginx.start "1"
+Nginx.stop
+Nginx.start "4 loglevel=warn"
 Minitest.after_run do
   Nginx.stop
 end
@@ -114,11 +122,17 @@ class UpstreamTest <  Minitest::Test
     @upstreams.each { |up| up.stop }
   end
   class Upstream
-    DEFAULT_WEIGHT = 100
+    DEFAULT_WEIGHT = 1
     def hit(srv_name)
       @hits[srv_name] = (@hits[srv_name] || 0) + 1
     end
     attr_accessor :hits, :weights, :name
+    
+    def response_counts_match?
+      return true if not @responses
+      total_hits = @hits.values.sum.to_f || 0
+      total_hits == @responses[:ok]
+    end
     
     def balanced?(max_error=0.05)
       total_weight = @weights.values.sum.to_f || 0
@@ -129,38 +143,53 @@ class UpstreamTest <  Minitest::Test
         errors[name]=((@hits[name] || 0) - expected)/expected
       end
       if errors.values.max.abs > max_error
-        msg = errors.map {|k, v| "#{k}:#{v>0 ? '+':'-'}#{v.abs*100}%"}.join ", "
+        msg = errors.map {|k, v| "#{k}:#{v>0 ? '+':'-'}#{(v.abs*100).round}%"}.join ", "
         return false, msg
       end
       return true
     end
     
-    def initialize(name, servers={})
+    def initialize(name, servers=[], weights=nil)
       @name = name
       @hits = {}
       @weights = {}
       @servers = {}
-      servers.each do |server_config|
+      servers.each_with_index do |server_config, i|
         server = start_server(server_config) do |env, this_server|
           path = env["REQUEST_PATH"] || env["PATH_INFO"]
-          puts "hit path #{path}"
           if path != "/ready"
             self.hit this_server.name
           end
         end
         @servers[server.name] = server
-        weight = Hash === server_config ? server_config[:weight] : DEFAULT_WEIGHT
-        @weights[server.name] = weight || DEFAULT_WEIGHT
+        @weights[server.name] = ((Hash === server_config) && server_config[:weight]) || (weights && weights[i]) || DEFAULT_WEIGHT
       end
     end
     
     def stop
-      name, srv = @servers.first
-      srv.stop if srv #thanks to a celluloid quirk, this stops all supervised servers
-      @servers.each do |name, v|
+      @servers.each do |name, srv|
+        srv.stop
         get_until "http://#{name}/ready", :couldnt_connect
       end
       @servers = {}
+    end
+    
+    def server(name)
+      if Numeric === name
+        srv_name, srv = @servers.find {|k, v| v.port == name}
+        return srv
+      end
+      return srv[name]
+    end
+    
+    def request(url=nil, repeat=1000, opt={})
+      url ||= "/#{@name}"
+      @responses ||= {}
+      resps = get_repeat(url, repeat, opt)
+      resps.each do |k,v|
+        @responses[k] = (@responses[k] || 0) + v
+      end
+      resps
     end
     
     private
@@ -177,21 +206,28 @@ class UpstreamTest <  Minitest::Test
     end
   end
     
-  def upstream(name, opt)
-    up = Upstream.new name, opt
+  def upstream(name, servers, weights=nil)
+    up = Upstream.new name, servers, weights
     @upstreams << up
     up
   end
   
   def assert_balanced(upstream, max_error=0.05)
+    assert upstream.response_counts_match?
     ok, err = upstream.balanced?(max_error)
     err = "upstream #{upstream.name} not balanced: #{err}" if not ok
     assert ok, err
   end
   
   def test_simple_roundrobin
-    up =  upstream "simple_roundrobin", [8083, 8084]
-    10.times { get "/simple_roundrobin"}
+    up =  upstream "simple_roundrobin", [8083, 8084, 8085]
+    up.request
+    assert_balanced up
+  end
+  
+  def test_weighted_roundrobin
+    up =  upstream "weighted_roundrobin", [8083, 8084, 8085], [1, 10, 30]
+    up.request
     assert_balanced up
   end
 end
